Inter-server handshake mechanism is implemented
The communication channel between the servers is now more secure: the server can no longer connect to the system without the correct access key.

Authorisation is implemented by a handshake mechanism. When a server is connected to a hub, a random 512-bit key is sent to it. The server uses built-in V8 cryptographic functions to generate a sequence using the password given in the configuration and the key received from the hub. The server then sends the resulting sequence to the hub, which generates its own sequence in the same way. The sequences will match if the access keys on the server and the hub match. The password itself is not transmitted over the network and cannot be intercepted.