Refactoring checkpoint
I'm a craze, so I decided to rewrite the Blessed module with TypeScript so that the project doesn't have any compilation errors. Thanks to TypeScript for generating code even if there are errors in it. So far there are 1904 of them, 1898 of which come from the Blessed library. Nevertheless, this can be considered a success, as there were originally about 2700 errors. This commit is just the intermediate result of (another) great refactoring.
Despite all the chaos in the code now, the server compiles and runs, which makes me very happy (and even a little surprised).