Continuation of the Great Refactoring
Gradually getting rid of the old-fashioned code that was used in the module. For example, __defineGetter__ and __defineSetter__. Implicit getters and setters are bad practice in my opinion. The programmer can't know when he's changing a variable and when he's calling something.
At the moment the server is running in 20 by 60 mode. Scrolling is temporarily broken, I guess because I removed the part of the code responsible for automatic layout. Basically, I remove everything from the library that I don't use (and that I don't feel sorry for). The code previously responsible for shrinking was in the same place as the keyboard and mouse handlers, and lots of other code. I'm just trying to structurize the code.
There are 1381 TypeScript compiler errors left. Almost half are defeated.