Upgrading the "Blessed" module
I probably should have started a separate repository to work on this library, but I'm too lazy to do that.
So, I halved the number of compilation errors again. There are 673 of them now.
While working on this library I came across a lot of strange code. In the main class of the library, I found several dozen methods that were copied two to four times along with all the code, and the copies differed only in name. Some of them didn't seemed to be even used anywhere. Abbreviations, labels, overcomplicated constructs everywhere... No, I'm not judging anyone, I wrote such code myself. For me such refactoring is now a challenge for myself. Will I manage to get to 0 errors? How beautiful will the code be? I cannot wait to see the clean, beautiful code where everything is in its place.
As usual, I got carried away talking off-topic. As a matter of fact, I've removed all the renamed copies of methods (which reduced the file by 700+ lines). I have removed the check that the Node is attached to the Screen immediately. This makes no sense, because the Node doesn't have to initialize completely until it's visible.
I've removed all the code associated with moving elements with the mouse, as well as the code responsible for scrollbars and labels. The scrollbars made the position counting code too complicated. In general they should be separate elements with their own logic. Their presence should not be checked in ifs everywhere as it was before. If I need scrollbars, I will reimplement them.
Oh, and I also deleted some useless files that happened to be in one of the past commit.
One of the tasks for the next commit is to get rid of the "type" field, which stores the element type as a string and whose value is checked in ifs in unexpected places.