Spreading the library code by file
So, I've put the code responsible for redrawing the screen into the "Framebuffer" class. I also removed all uses of "lpos" and "_getCoords", because now since automatic layout has been removed from the library, these functions are meaningless.

I was surprised to find that the code responsible for indents and borders continues to work correctly. I have added a corresponding test.

There are 40 errors left in the project. I think they will resolve themselves in the course of structuring the library.