Completing refactoring
I've done almost everything I wanted to do with the Blessed library. I fixed the tab handling, made edits to the event handling, and merged two EventEmitter classes into one. (I had one from the beginning, the other was inherited from the library, it has richer functionality so I will use it from now on). There are 28 TypeScript compilation errors left, but they should fix themselves in time as the code is structured. The huge painful refactoring is over. For this event, I even moved all the console developments to the server code and added a gray text tooltip to the console. It appears when only one autocomplete option is possible. Very handy feature. I really like my console. It looks so... legacy