Cleaned up the code responsible for connections/clients
There used to be a lot of confusion in the code: what "client" and "connection" meant. To be honest, I didn't figure it out myself, so I came up with a new ideology and brought the whole code to it: connection deals with sending data, client deals with its handling. It's also now possible to pipe connections together to form pipelines. For example, you can pipe another connection, which will be responsible for handling authorization or handshaking to a websocket connection. This was useful for embedded server implementation (where you don't need to have sockets at all, two different connections are just short-circuited to each other) and cluster connections (which I haven't even tested yet. I'll do it later)