World collider is now a world entity component.
Game crashes when you shoot for some reason. It does not give you any error information, because Blessed captures this output and then closes the window. I'm going to switch to lightweight 'serverline' module in order to fix this. (There are actually a million reasons why I've long wanted to replace this clumping lump of spaghetti code with something that would solve the asynchronous I/O problem well)