Replaced old BinaryEncoder and BinaryDecoder with proper buffers
New ReadBuffer and WriteBuffer classes are based on native DataView class, which is a lot faster and uses single buffer for all the data. It's now possible to easily implement "block-based" binary frame, where each block could be either processed or skipped. In the previous buffer implementation it would require a lot of offsets to be stored in the frame. On the way of making perfect ECS-based game architecture, it was very important to take this step.

Also, all existent maps was converted to new file format. This was done by creating a temporary ugly script (and a separate build scheme). This script will be deleted from the source files shortly, as well as old BinaryEncoder/BinaryDecoder classes.