Removing old packet system, messaging system is now capable of doing the entity tree synchronisation.
A nightmare. The current state of the game cannot be described by any different word. The game is capable of doing the single thing, and it's to create a tank on the map by using the new messaging system. Everything else is broken. I'd love to fix it, especially when it becomes easier and easier to do so, but... It's 1.30 AM and I want to sleep very bad.