Started to add some instruments
Implemented toolbar and some instruments: pencil, eraser, move, scale and fill. It's not possible to change the brush size or pick the block just yet. Swapped the y direction of the game, removed the minus in Math.atan2(-y, x) where I could. Yet to test, but it feels much cleaner now.